[flexiglass] Verify that NSSL#mIntrinsicContentHeight is not accessed
NSSL#mIntrinsicContentHeight is replaced by ScrollViewFields#.intrinsicStackHeight when flexiglass is enabled.

Bug: 296118689
Test: run sysui with, and without flexiglass
Flag: EXEMPT mechanical refactor

+++ packages/SystemUI/src/com/android/systemui/statusbar/notification/stack/NotificationStackScrollLayoutController.java
@@ -1188,7 +1188,7 @@ public class NotificationStackScrollLayoutController implements Dumpable {
 
     public int getIntrinsicContentHeight() {
         SceneContainerFlag.assertInLegacyMode();
-        return mView.getIntrinsicContentHeight();
+        return (int) mView.getIntrinsicContentHeight();
     }
 
     /**
